requirements.txt
Flask~=2.0
PyMySQL~=1.0
python-dotenv
